svn:2877 Clean up some bugs in mkendf.sh and add mkfils.py, which creates the "plus" and "minus" files used in assimilation.

+++ util/mkendf/mkendf.sh
@@ -1,53 +1,68 @@
 #!/bin/bash
-echo ' Processing Empire output'
-echo ''
 
 # par 1 = working directory
-# par 2 = project (file) name (without extension)
+# par 2 = project (file) name
 
-cd $1
+fil=${2%.*}
 
-matnum=$(getmtn $2_orig.endf)
+if [ ! -e ${fil}.res ]
+then
+    echo ' Resonance file not found!'
+    echo ' Please create file '${fil}.res
+    exit
+fi
+
+matnum=$(getmtn ${fil}.res)
 matnum=${matnum:?" undefined!"}
 echo 'MAT = '$matnum
 
+cd $1
+
+echo ' Processing Empire output'
+echo ''
 # remove any old temp files
 
-if [ -e $2.endf ]
+if [ -e ${fil}.endf ]
 then
-        rm -r $2.endf
+    rm -r ${fil}.endf
 fi
 
-if [ -e $2_1.endf ]
+if [ -e ${fil}_1.endf ]
 then
-        rm -r $2_1.endf
+    rm -r ${fil}_1.endf
 fi
 
-if [ -e $2_2.endf ]
+if [ -e ${fil}_2.endf ]
 then
-        rm -r $2_2.endf
+    rm -r ${fil}_2.endf
 fi
 
-if [ -e $2_2.endfadd ]
+if [ -e ${fil}_2.endfadd ]
 then
-        rm -r $2_2.endfadd
+    rm -r ${fil}_2.endfadd
 fi
 
-if [ -e $2_3.endf ]
+if [ -e ${fil}_3.endf ]
 then
-        rm -r $2_3.endf
+    rm -r ${fil}_3.endf
 fi
 
-if [ -e $2.njoy ]
+if [ -e ${fil}.njoy ]
 then
-        rm -r $2.njoy
+    rm -r ${fil}.njoy
 fi
 
 # convert Empire output to ENDF format & add resonances
 
+if [ ! -e ${fil}.out ]
+then
+    echo ' Empire output file not found!'
+    exit
+fi
+
 cat > input <<EOF
-$2.out
-$2_1.endf
+${fil}.out
+${fil}_1.endf
 1
 -1.0
 $matnum
@@ -56,33 +71,35 @@ EOF
 /home/herman/empire/util/empend/empend < input
 
 cat > input <<EOF
-$2_1.endf
-$2.res
-$2_2.endf
+${fil}_1.endf
+../${fil}.res
+${fil}_2.endf
 0
 EOF
 
 /home/herman/empire/util/endres/endres < input
 
-rm $2_1.endf
+rm ${fil}_1.endf
 rm empmf1.tmp
 rm endres.scr
 rm angdis.*
 rm input
 
 # add in nubar, fission spectra(MF5,MT18) and
-# ang dist (MF4,MT18) from file $2_orig.endf
+# ang dist (MF4,MT18) from file ${fil}_orig.endf
+# only do this if it's a fissile material!
 
-/home/herman/empire/util/mkendf/add_endf $2_2.endf $2_orig.endf
-rm $2_2.endf
+/home/herman/empire/util/mkendf/add_endf ${fil}_2.endf ../${fil}_orig.endf
+rm ${fil}_2.endf
+# cp ${fil}_2.endf ${fil}_2.endfadd
 
 # add 103-107 if data available and re-make elastic
 # FIXUP reads from local file, NOT input stream
 
 cat > FIXUP.INP <<EOF
 10101111111001
-$2_2.endfadd
-$2_3.endf
+${fil}_2.endfadd
+${fil}_3.endf
 S103=+(600,649)
 S104=+(650,699)
 S105=+(700,749)
@@ -105,21 +122,20 @@ EOF
 
 /home/herman/empire/util/fixup/fixup
 echo ' Done with FIXUP'
-rm $2_2.endfadd
+rm ${fil}_2.endfadd
 rm FIXUP.INP
 echo ' Done cleaning FIXUP'
 
 # run STAN
 
-/home/herman/empire/util/stan/stan $2_3.endf
-mv $2_3.STN $2.endf
-rm $2_3.endf
+/home/herman/empire/util/stan/stan ${fil}_3.endf
+mv ${fil}_3.STN ${fil}.endf
+rm ${fil}_3.endf
 
 # finally, process this file with NJOY
-# runNJOY.py ../njoyActinides.sh $2.endf
-# FIX NJOY scripts to use local working directories!!
+runNJOY.py ../njoy_33grp.sh ${fil}.endf
 
-# To run NJOY on local machine:
-# ../njoyActinides.sh $2.endf
+# To run NJOY on local machine (don't!):
+# ../njoyActinides.sh ${fil}.endf
 
 exit
